requirements.txt
rdflib
rdflib-jsonld
pyramid
waitress
# icc.rdfservice
# icc.restfuldocs
pyramid_debugtoolbar
pyramid_chameleon
Babel
lingua
bcrypt
cryptography
sendgrid
mammoth
#secretary
pyramid_tm
sqlalchemy
sqlalchemy-utils
mmh3
zope.sqlalchemy
pyramid_sacrud
# ps_alchemy
# https://github.com/isu-enterprise/ps_alchemy/archive/master.zip#egg=ps_alchemy
-e git+https://github.com/isu-enterprise/ps_alchemy#egg=ps_alchemy
phonenumbers
requests
# psycopg2-binary
https://github.com/psycopg/psycopg2/archive/2_7_6.tar.gz



-e git+https://github.com/isu-enterprise/isu.enterprise#egg=isu.enterprise
-e git+https://github.com/isu-enterprise/isu.webapp#egg=isu.webapp
-e git+https://github.com/isu-enterprise/isu.onece#egg=isu.onece
-e git+https://github.com/isu-enterprise/icc.mvw#egg=icc.mvw
-e git+https://github.com/isu-enterprise/pyramid_storage#egg=pyramid_storage
# pyramid_elfinder
-e git+https://github.com/ITCase/pyramid_elfinder.git#egg=pyramid_elfinder
https://github.com/eugeneai/docx2html/archive/v0.2.4.tar.gz#egg=docx2html
lxml
#xlrd
#pandas
#marisa-trie
#pymorphy2
pyramid_zcml
waitress
